requirements.txt
acvl_utils==0.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens 
async-lru==2.0.4
asyncio-atexit==1.0.1
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
backoff==2.2.1
batchgenerators==0.25
batchgeneratorsv2==0.2
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
comm 
connected-components-3d==3.18.0
contourpy==1.2.1
cycler==0.12.1
debugpy 
decorator 
defusedxml==0.7.1
Deprecated==1.2.14
dicom2nifti==2.4.11
docker-pycreds==0.4.0
docopt==0.6.2
dynamic_network_architectures==0.3.1
einops==0.8.0
exceptiongroup 
executing 
fastjsonschema==2.20.0
fft-conv-pytorch==1.2.0
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.63.2
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
imagecodecs==2024.6.1
imageio==2.34.2
importlib-metadata==6.11.0
iopath==0.1.10
ipykernel 
ipympl==0.9.4
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi 
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client 
jupyter_core 
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
linecache2==1.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline 
mistune==3.0.2
monai==1.3.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio 
networkx==3.3
nibabel==5.2.1
nnunetv2==2.5.1
notebook_shim==0.2.4
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
overrides==7.7.0
packaging 
pandas==2.2.2
pandocfilters==1.5.1
parso 
pexpect 
pickleshare 
pillow==10.4.0
pipreqs==0.5.0
platformdirs 
portalocker==2.10.1
prometheus_client==0.20.0
prompt_toolkit 
protobuf==4.25.4
psutil==5.9.8
ptyprocess 
pure_eval 
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydicom==2.4.4
Pygments 
pyparsing==3.1.2
python-dateutil 
python-gdcm==3.0.24.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq 
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.26.0
setproctitle==1.3.5
setuptools==69.5.1
SimpleITK==2.3.1
six 
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
stack-data 
sympy==1.13.3
synapseclient==4.4.0
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.24
tinycss2==1.3.0
tornado 
tqdm==4.66.4
traceback2==1.4.0
traitlets 
triton==3.0.0
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions 
tzdata==2024.1
unittest2==1.1.0
uri-template==1.3.0
urllib3==1.26.19
wandb==0.19.9
wcwidth 
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
yacs==0.1.8
yarg==0.1.9
zipp 
